refactor: move Todo components to lib/components, update page structure, and add new routes
