+++ src/routes/todo/+page.svelte
@@ -1,8 +1,8 @@
 <script  lang="ts">
 	import type { Todo } from '../../types/type';
-	import TodoForm from './TodoForm.svelte';
-	import TodoHeader from './TodoHeader.svelte';
-	import TodoList from './TodoList.svelte';
+	import TodoForm from '../../lib/components/TodoForm.svelte';
+	import TodoHeader from '../../lib/components/TodoHeader.svelte';
+	import TodoList from '../../lib/components/TodoList.svelte';
 
 	
 	let list: Todo[] = $state([]);
